pyproject.toml
[project]
name = "raitap"
dynamic = ["version"]
authors = [
    { name = "Philipp Denzel", email = "denp@zhaw.ch" },
    { name = "Oliver Forster", email = "foro@zhaw.ch" },
    # add your name / email here
]
maintainers = [
    { name = "Philipp Denzel", email = "denp@zhaw.ch" },
    { name = "Oliver Forster", email = "foro@zhaw.ch" },
    { name = "Stanislas Laurent", email = "lauresta@students.zhaw.ch" },
    { name = "Jonas Vonderhagen", email = "vondejon@students.zhaw.ch"}
]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
keywords = ["artificial intelligence", "deep learning", "neural networks", "technical assessment", "responsible AI"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core framework
    "hydra-core>=1.3.2",
    "mlflow>=3.9.0",
    "numpy>=2.4.2",
    # Deep learning (vision)
    "torch>=2.10.0",
    "torchvision>=0.20.0",
    "pillow>=10.0.0",
    # Transparency/explainability
    "shap>=0.46.0",
    "captum>=0.7.0",
    # Visualization
    "matplotlib>=3.9.0",
]

[project.scripts]
raitap = "raitap.run:main"

[project.urls]
Documentation = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform#readme"
Homepage = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform"
Repository = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform.git"
Issues = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform/issues"
Source = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform"

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "commitizen>=3.0.0",
    "pyright>=1.1.390",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/raitap/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/raitap"]
include = [
    "src/raitap/**/*.yaml",
    "src/raitap/**/*.yml",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true

[tool.pyright]
include = ["src"]
exclude = [
    "**/.*",
    "**/.venv",
    "**/__pycache__",
    "**/node_modules",
    "**/build",
    "**/dist",
]
pythonVersion = "3.13"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
